+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Rows/Extends/TimeTags/TimeTagEditRowExtend.cs
@@ -5,13 +5,13 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Game.Rulesets.Karaoke.Objects;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Extends.Notes
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Extends.TimeTags
 {
-    public class NoteExtend : RowEditExtend
+    public class TimeTagEditRowExtend : EditRowExtend
     {
-        public override float ContentHeight => 180;
+        public override float ContentHeight => 100;
 
-        public NoteExtend(Lyric lyric)
+        public TimeTagEditRowExtend(Lyric lyric)
             : base(lyric)
         {
         }
@@ -24,10 +24,11 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Extends.Notes
 
         protected override Drawable CreateContent(Lyric lyric)
         {
-            return new NoteEditor(lyric)
+            // todo : waiting for implementation.
+            return new TimeTagEditor(lyric)
             {
                 RelativeSizeAxes = Axes.X,
-                Height = 150,
+                Height = 100,
             };
         }
     }
